fix special characters and broken links: roffit
Fixed one tabbed indent, and tried to implement corrections in "text2name()" and "do_encode()" again. Results - this is the best I'm going to do. Was not able to get working in "do_encode()" at all. In "text2name()" had very minor unexpected results in one file "brew.1", and was not able to resolve. It reproduced multiple links of the same link on several lines in that file.